+++ web_app/MediaKraken/admins/__init__.py
@@ -1,3 +1,13 @@
 # -*- coding: utf-8 -*-
 """The user module."""
 from . import views
+from . import views_backup
+from . import views_chromecasts
+from . import views_cron
+from . import views_docker
+from . import views_library
+from . import views_link
+from . import views_share
+from . import views_transmission
+from . import views_tvtuners
+from . import views_users
