+++ web_app/MediaKraken/admins/views.py
@@ -19,14 +19,10 @@ import flask
 from flask_login import current_user
 from functools import wraps
 from functools import partial
-from MediaKraken.admins.forms import BookAddForm
-from MediaKraken.admins.forms import CronEditForm
-from MediaKraken.admins.forms import LibraryAddEditForm
-from MediaKraken.admins.forms import ShareAddEditForm
-from MediaKraken.admins.forms import BackupEditForm
-from MediaKraken.admins.forms import DLNAEditForm
-from MediaKraken.admins.forms import UserEditForm
 from MediaKraken.admins.forms import AdminSettingsForm
+from MediaKraken.admins.forms import BackupEditForm
+from MediaKraken.admins.forms import BookAddForm
+from MediaKraken.admins.forms import DLNAEditForm
 
 from common import common_config_ini
 from common import common_internationalization
@@ -38,7 +34,6 @@ from common import common_network
 from common import common_pagination
 from common import common_string
 from common import common_system
-from common import common_transmission
 from common import common_version
 from common import common_zfs
 import database as database_base
@@ -150,118 +145,12 @@ def admins():
                           )
 
 
-@blueprint.route("/users")
-@blueprint.route("/users/")
-@login_required
-@admin_required
-def admin_users():
-    """
-    Display user list
-    """
-    page, per_page, offset = common_pagination.get_page_items()
-    pagination = common_pagination.get_pagination(page=page,
-                                                  per_page=per_page,
-                                                  total=g.db_connection.db_user_list_name_count(),
-                                                  record_name='users',
-                                                  format_total=True,
-                                                  format_number=True,
-                                                 )
-    return render_template('admin/admin_users.html',
-                           users=g.db_connection.db_user_list_name(offset, per_page),
-                           page=page,
-                           per_page=per_page,
-                           pagination=pagination,
-                          )
-
-
-@blueprint.route("/user_detail/<guid>/")
-@blueprint.route("/user_detail/<guid>")
-@login_required
-@admin_required
-def admin_user_detail(guid):
-    """
-    Display user details
-    """
-    return render_template('admin/admin_user_detail.html',
-        data_user=g.db_connection.db_user_detail(guid))
-
-
 #@blueprint.route("/dlna")
 #@blueprint.route("/dlna/")
 #@login_required
 #@admin_required
 #def admin_dlna():
 #    return render_template("admin/admin_dlna.html")
-
-
-@blueprint.route('/cron')
-@blueprint.route('/cron/')
-@login_required
-@admin_required
-def admin_cron_display_all():
-    """
-    Display cron jobs
-    """
-    page, per_page, offset = common_pagination.get_page_items()
-    pagination = common_pagination.get_pagination(page=page,
-                                                  per_page=per_page,
-                                                  total=g.db_connection.db_cron_list_count(False),
-                                                  record_name='Cron Jobs',
-                                                  format_total=True,
-                                                  format_number=True,
-                                                 )
-    return render_template('admin/admin_cron.html',
-                           media_cron=g.db_connection.db_cron_list(False, offset, per_page),
-                           page=page,
-                           per_page=per_page,
-                           pagination=pagination,
-                          )
-
-
-@blueprint.route('/cron_edit/<guid>/', methods=['GET', 'POST'])
-@blueprint.route('/cron_edit/<guid>', methods=['GET', 'POST'])
-@login_required
-@admin_required
-def admin_cron_edit(guid):
-    """
-    Edit cron job page
-    """
-    form = CronEditForm(request.form, csrf_enabled=False)
-    if form.validate_on_submit():
-        pass
-    return render_template('admin/admin_cron_edit.html', guid=guid, form=form)
-
-
-@blueprint.route("/chromecasts", methods=["GET", "POST"])
-@blueprint.route("/chromecasts/", methods=["GET", "POST"])
-@login_required
-@admin_required
-def admin_chromecasts():
-    """
-    List chromecasts
-    """
-    device_list = []
-    for row_data in g.db_connection.db_device_list('cast'):
-        device_list.append((row_data['mm_device_id'], row_data['mm_device_json']['Name'],
-                            row_data['mm_device_json']['Model'],
-                            row_data['mm_device_json']['IP'], True))
-    return render_template("admin/admin_chromecasts.html", data_chromecast=device_list)
-
-
-@blueprint.route("/tvtuners", methods=["GET", "POST"])
-@blueprint.route("/tvtuners/", methods=["GET", "POST"])
-@login_required
-@admin_required
-def admin_tvtuners():
-    """
-    List tvtuners
-    """
-    tv_tuners = []
-    for row_data in g.db_connection.db_tuner_list():
-        tv_tuners.append((row_data['mm_tuner_id'], row_data['mm_tuner_json']['HWModel']\
-        + " (" + row_data['mm_tuner_json']['Model'] + ")", row_data['mm_tuner_json']['IP'],
-        row_data['mm_tuner_json']['Active'], len(row_data['mm_tuner_json']['Channels'])))
-    return render_template("admin/admin_tvtuners.html", data_tuners=tv_tuners)
 
 
 @blueprint.route("/nas", methods=["GET", "POST"])
@@ -274,49 +163,6 @@ def admin_nas():
     """
     nas_devices = []
     return render_template("admin/admin_nas.html", data_nas=nas_devices)
-
-
-@blueprint.route("/transmission")
-@blueprint.route("/transmission/")
-@login_required
-@admin_required
-def admin_transmission():
-    """
-    Display transmission page
-    """
-    trans_connection = common_transmission.CommonTransmission(option_config_json)
-    transmission_data = []
-    if trans_connection is not None:
-        torrent_no = 1
-        for torrent in trans_connection.com_trans_get_torrent_list():
-            transmission_data.append((common_internationalization.com_inter_number_format(torrent_no),
-                torrent.name, torrent.hashString, torrent.status, torrent.progress, torrent.ratio))
-            torrent_no += 1
-    return render_template("admin/admin_transmission.html", data_transmission=transmission_data)
-
-
-@blueprint.route('/transmission_delete', methods=["POST"])
-@login_required
-@admin_required
-def admin_transmission_delete_page():
-    """
-    Delete torrent from transmission
-    """
-    #g.db_connection.db_Audit_Path_Delete(request.form['id'])
-    #g.db_connection.db_commit()
-    return json.dumps({'status': 'OK'})
-
-
-@blueprint.route('/transmission_edit', methods=["POST"])
-@login_required
-@admin_required
-def admin_transmission_edit_page():
-    """
-    Edit a torrent from transmission
-    """
-    #g.db_connection.db_Audit_Path_Delete(request.form['id'])
-    #g.db_connection.db_commit()
-    return json.dumps({'status': 'OK'})
 
 
 @blueprint.route('/books_add', methods=['GET', 'POST'])
@@ -345,340 +191,6 @@ def admin_books_add():
     return render_template("admin/admin_books_add.html", form=form)
 
 
-@blueprint.route("/library", methods=["GET", "POST"])
-@blueprint.route("/library/", methods=["GET", "POST"])
-@login_required
-@admin_required
-def admin_library():
-    """
-    List all media libraries
-    """
-    if request.method == 'POST':
-        g.db_connection.db_trigger_insert(('python', './subprogram_file_scan.py'))
-        flash("Scheduled media scan.")
-    page, per_page, offset = common_pagination.get_page_items()
-    pagination = common_pagination.get_pagination(page=page,
-                                                  per_page=per_page,
-                                                  total=g.db_connection.db_table_count(
-                                                      'mm_media_dir'),
-                                                  record_name='library dir(s)',
-                                                  format_total=True,
-                                                  format_number=True,
-                                                 )
-    return render_template("admin/admin_library.html",
-                           media_dir=g.db_connection.db_audit_paths(offset, per_page),
-                           page=page,
-                           per_page=per_page,
-                           pagination=pagination,
-                          )
-
-
-@blueprint.route("/library_edit", methods=["GET", "POST"])
-@blueprint.route("/library_edit/", methods=["GET", "POST"])
-@login_required
-@admin_required
-def admin_library_edit_page():
-    """
-    allow user to edit lib
-    """
-    form = LibraryAddEditForm(request.form)
-    if request.method == 'POST':
-        if form.validate_on_submit():
-            if request.form['action_type'] == 'Add':
-                # check for UNC
-                if request.form['library_path'][:1] == "\\":
-                    addr, share, path = common_string.com_string_unc_to_addr_path(
-                        request.form['library_path'])
-                    logging.info('smb info: %s %s %s', (addr, share, path))
-                    if addr is None: # total junk path for UNC
-                        flash("Invalid UNC path.", 'error')
-                        return redirect(url_for('admins.admin_library_edit_page'))
-                    smb_stuff = common_network_cifs.CommonCIFSShare()
-                    smb_stuff.com_cifs_connect(addr)
-                    if not smb_stuff.com_cifs_share_directory_check(share, path):
-                        smb_stuff.com_cifs_close()
-                        flash("Invalid UNC path.", 'error')
-                        return redirect(url_for('admins.admin_library_edit_page'))
-                    smb_stuff.com_cifs_close()
-                # smb/cifs mounts
-                elif request.form['library_path'][0:3] == "smb":
-                    # TODO
-                    smb_stuff = common_network_cifs.CommonCIFSShare()
-                    smb_stuff.com_cifs_connect(ip_addr, user_name='guest', user_password='')
-                    smb_stuff.com_cifs_share_directory_check(share_name, dir_path)
-                    smb_stuff.com_cifs_close()
-                # nfs mount
-                elif request.form['library_path'][0:3] == "nfs":
-                    # TODO
-                    pass
-                elif not os.path.isdir(request.form['library_path']):
-                    flash("Invalid library path.", 'error')
-                    return redirect(url_for('admins.admin_library_edit_page'))
-                # verify it doesn't exit and add
-                if g.db_connection.db_audit_path_check(request.form['library_path']) == 0:
-                    g.db_connection.db_audit_path_add(request.form['library_path'],
-                        request.form['Lib_Class'], request.form['Lib_Share'])
-                    g.db_connection.db_commit()
-                    return redirect(url_for('admins.admin_library'))
-                else:
-                    flash("Path already in library.", 'error')
-                    return redirect(url_for('admins.admin_library_edit_page'))
-            elif request.form['action_type'] == 'Browse...': # popup browse form
-                pass
-            elif request.form['action_type'] == 'Synology': # popup browse form for synology
-                pass
-        else:
-            flash_errors(form)
-    class_list = []
-    for row_data in g.db_connection.db_media_class_list():
-        if row_data['mm_media_class_display']: # flagged for display
-            class_list.append((row_data['mm_media_class_type'], row_data['mm_media_class_guid']))
-    share_list = []
-    for row_data in g.db_connection.db_audit_shares():
-        share_name = row_data['mm_media_share_server'] + ":" + row_data['mm_media_share_path']
-        share_list.append((share_name, row_data['mm_media_share_guid']))
-
-    return render_template("admin/admin_library_edit.html", form=form,
-                           data_class=class_list,
-                           data_share=share_list)
-
-
-@blueprint.route('/library_delete', methods=["POST"])
-@login_required
-@admin_required
-def admin_library_delete_page():
-    """
-    Delete library action 'page'
-    """
-    g.db_connection.db_audit_path_delete(request.form['id'])
-    g.db_connection.db_commit()
-    return json.dumps({'status': 'OK'})
-
-
-@blueprint.route('/getLibraryById', methods=['POST'])
-@login_required
-@admin_required
-def getLibraryById():
-    result = g.db_connection.db_audit_path_by_uuid(request.form['id'])
-    return json.dumps({'Id': result['mm_media_dir_guid'],
-        'Path': result['mm_media_dir_path'], 'Media Class': result['mm_media_dir_class_type']})
-
-
-@blueprint.route('/updateLibrary', methods=['POST'])
-@login_required
-@admin_required
-def updateLibrary():
-    g.db_connection.db_audit_path_update_by_uuid(request.form['new_path'],
-        request.form['new_class'], request.form['id'])
-    return json.dumps({'status': 'OK'})
-
-
-@blueprint.route("/share", methods=["GET", "POST"])
-@blueprint.route("/share/", methods=["GET", "POST"])
-@login_required
-@admin_required
-def admin_share():
-    """
-    List all share/mounts
-    """
-    page, per_page, offset = common_pagination.get_page_items()
-    pagination = common_pagination.get_pagination(page=page,
-                                                  per_page=per_page,
-                                                  total=g.db_connection.db_table_count(
-                                                      'mm_media_share'),
-                                                  record_name='share(s)',
-                                                  format_total=True,
-                                                  format_number=True,
-                                                 )
-    return render_template("admin/admin_share.html",
-                           media_dir=g.db_connection.db_audit_shares(offset, per_page),
-                           page=page,
-                           per_page=per_page,
-                           pagination=pagination,
-                          )
-
-
-@blueprint.route("/share_edit", methods=["GET", "POST"])
-@blueprint.route("/share_edit/", methods=["GET", "POST"])
-@login_required
-@admin_required
-def admin_share_edit_page():
-    """
-    allow user to edit share
-    """
-    form = ShareAddEditForm(request.form)
-    logging.info('hereeditshare')
-    if request.method == 'POST':
-        logging.info('herepost')
-        if form.validate_on_submit():
-            logging.info('action %s', request.form['action_type'])
-            if request.form['action_type'] == 'Add':
-                logging.info('type %s', request.form['storage_mount_type'])
-                # check for UNC
-                if request.form['storage_mount_type'] == "unc":
-#                    addr, share, path = common_string.com_string_unc_to_addr_path(\
-#                        request.form['storage_mount_path'])
-#                    if addr is None: # total junk path for UNC
-#                        flash("Invalid UNC path.", 'error')
-#                        return redirect(url_for('admins.admin_share_edit_page'))
-#                    logging.info('unc info: %s %s %s' % (addr, share, path))
-                    smb_stuff = common_network_cifs.CommonCIFSShare()
-                    smb_stuff.com_cifs_connect(request.form['storage_mount_server'],
-                        user_name=request.form['storage_mount_user'],
-                        user_password=request.form['storage_mount_password'])
-#                    if not smb_stuff.com_cifs_share_directory_check(\
-#                            request.form['storage_mount_server'], \
-#                            request.form['storage_mount_path']):
-#                        smb_stuff.com_cifs_close()
-#                        flash("Invalid UNC path.", 'error')
-#                        return redirect(url_for('admins.admin_share_edit_page'))
-                    smb_stuff.com_cifs_close()
-                # smb/cifs mounts
-                elif request.form['storage_mount_type'] == "smb":
-                    # TODO
-                    smb_stuff = common_network_cifs.CommonCIFSShare()
-                    smb_stuff.com_cifs_connect(request.form['storage_mount_server'],
-                        user_name=request.form['storage_mount_user'],
-                        user_password=request.form['storage_mount_password'])
-                    smb_stuff.com_cifs_share_directory_check(share_name,
-                        request.form['storage_mount_path'])
-                    smb_stuff.com_cifs_close()
-                # nfs mount
-                elif request.form['storage_mount_type'] == "nfs":
-                    # TODO
-                    pass
-                elif not os.path.isdir(request.form['storage_mount_path']):
-                    flash("Invalid share path.", 'error')
-                    return redirect(url_for('admins.admin_share_edit_page'))
-                # verify it doesn't exit and add
-                if g.db_connection.db_audit_share_check(request.form['storage_mount_path']) == 0:
-                    g.db_connection.db_audit_share_add(request.form['storage_mount_type'],
-                                                       request.form['storage_mount_user'],
-                                                       request.form['storage_mount_password'],
-                                                       request.form['storage_mount_server'],
-                                                       request.form['storage_mount_path'])
-                    g.db_connection.db_commit()
-                    return redirect(url_for('admins.admin_share'))
-                else:
-                    flash("Share already mapped.", 'error')
-                    return redirect(url_for('admins.admin_share_edit_page'))
-            elif request.form['action_type'] == 'Browse':
-                # browse for smb shares on the host network
-                pass
-        else:
-            flash_errors(form)
-    return render_template("admin/admin_share_edit.html", form=form)
-
-
-@blueprint.route('/share_delete', methods=["POST"])
-@login_required
-@admin_required
-def admin_share_delete_page():
-    """
-    Delete share action 'page'
-    """
-    g.db_connection.db_audit_share_delete(request.form['id'])
-    g.db_connection.db_commit()
-    return json.dumps({'status': 'OK'})
-
-
-@blueprint.route('/getShareById', methods=['POST'])
-@login_required
-@admin_required
-def getShareById():
-    result = g.db_connection.db_audit_share_by_uuid(request.form['id'])
-    return json.dumps({'Id': result['mm_share_dir_guid'],
-        'Path': result['mm_share_dir_path']})
-
-
-@blueprint.route('/updateShare', methods=['POST'])
-@login_required
-@admin_required
-def updateShare():
-    g.db_connection.db_audit_share_update_by_uuid(request.form['new_share_type'],
-                                                  request.form['new_share_user'],
-                                                  request.form['new_share_password'],
-                                                  request.form['new_share_server'],
-                                                  request.form['new_share_path'],
-                                                  request.form['id'])
-    return json.dumps({'status': 'OK'})
-
-
-@blueprint.route('/user_delete', methods=["POST"])
-@login_required
-@admin_required
-def admin_user_delete_page():
-    """
-    Delete user action 'page'
-    """
-    g.db_connection.db_user_delete(request.form['id'])
-    g.db_connection.db_commit()
-    return json.dumps({'status': 'OK'})
-
-
-@blueprint.route('/backup_delete', methods=["POST"])
-@login_required
-@admin_required
-def admin_backup_delete_page():
-    """
-    Delete backup file action 'page'
-    """
-    file_path, file_type = request.form['id'].split('|')
-    if file_type == "Local":
-        os.remove(file_path)
-    elif file_type == "AWS" or file_type == "AWS S3":
-        CLOUD_HANDLE.com_cloud_file_delete('awss3', file_path, True)
-    return json.dumps({'status': 'OK'})
-
-
-@blueprint.route("/backup", methods=["GET", "POST"])
-@blueprint.route("/backup/", methods=["GET", "POST"])
-@login_required
-@admin_required
-def admin_backup():
-    """
-    List backups from local fs and cloud
-    """
-    form = BackupEditForm(request.form)
-    if request.method == 'POST':
-        if form.validate_on_submit():
-            if request.form['backup'] == 'Update':
-                pass
-            elif request.form['backup'] == 'Start Backup':
-                g.db_connection.db_trigger_insert(('python',
-                    './subprogram_postgresql_backup.py')) # this commits
-                flash("Postgresql Database Backup Task Submitted.")
-        else:
-            flash_errors(form)
-    backup_enabled = False
-    backup_files = []
-    for backup_local in common_file.com_file_dir_list(
-            option_config_json['MediaKrakenServer']['BackupLocal'], 'dump', False, False, True):
-        backup_files.append((backup_local[0], 'Local',
-            common_string.com_string_bytes2human(backup_local[1])))
-    # cloud backup list
-    for backup_cloud in CLOUD_HANDLE.com_cloud_backup_list():
-        backup_files.append((backup_cloud.name, backup_cloud.type,
-            common_string.com_string_bytes2human(backup_cloud.size)))
-    page, per_page, offset = common_pagination.get_page_items()
-    pagination = common_pagination.get_pagination(page=page,
-                                                  per_page=per_page,
-                                                  total=len(backup_files),
-                                                  record_name='backups',
-                                                  format_total=True,
-                                                  format_number=True,
-                                                 )
-    return render_template("admin/admin_backup.html", form=form,
-                           backup_list=sorted(backup_files, reverse=True),
-                           data_interval=('Hours', 'Days', 'Weekly'),
-                           data_class=common_cloud.CLOUD_BACKUP_CLASS,
-                           data_enabled=backup_enabled,
-                           page=page,
-                           per_page=per_page,
-                           pagination=pagination
-                          )
-
-
 @blueprint.route("/ffmpeg_stat")
 @blueprint.route("/ffmpeg_stat/")
 @login_required
@@ -688,27 +200,6 @@ def ffmpeg_stat():
     Redirect to ffmeg stats
     """
     return redirect("http://www.example.com/status.html", code=302)  # status.html is ffmpeg page
-
-
-@blueprint.route("/docker_stat")
-@blueprint.route("/docker_stat/")
-@login_required
-@admin_required
-def docker_stat():
-    """
-    Docker statistics including swarm
-    """
-    docker_inst = common_docker.CommonDocker()
-    # it returns a dict, not a json
-    docker_info = docker_inst.com_docker_info()
-    if docker_info['Swarm']['Managers'] == 0:
-        docker_swarm = "Cluster not active"
-    else:
-        docker_swarm = docker_inst.com_docker_swarm_inspect()['JoinTokens']['Worker']
-    return render_template("admin/admin_docker.html",
-                           data_host=docker_info,
-                           data_swam=docker_swarm
-                          )
 
 
 @blueprint.route("/server_stat")
@@ -762,41 +253,6 @@ def admin_server_zfs():
     """
     return render_template("admin/admin_server_zfs.html",
                            data_zpool=common_zfs.com_zfs_zpool_list())
-
-
-@blueprint.route("/link_server")
-@blueprint.route("/link_server/")
-@login_required
-@admin_required
-def admin_server_link_server():
-    """
-    Display page for linking server
-    """
-    page, per_page, offset = common_pagination.get_page_items()
-    pagination = common_pagination.get_pagination(page=page,
-                                                  per_page=per_page,
-                                                  total=g.db_connection.db_link_list_count(),
-                                                  record_name='linked servers',
-                                                  format_total=True,
-                                                  format_number=True,
-                                                 )
-    return render_template("admin/admin_link.html",
-                           data=g.db_connection.db_link_list(offset, per_page),
-                           page=page,
-                           per_page=per_page,
-                           pagination=pagination)
-
-
-@blueprint.route('/link_delete', methods=["POST"])
-@login_required
-@admin_required
-def admin_link_delete_page():
-    """
-    Delete linked server action 'page'
-    """
-    g.db_connection.db_link_delete(request.form['id'])
-    g.db_connection.db_commit()
-    return json.dumps({'status':'OK'})
 
 
 @blueprint.route("/cloud_google_drive")
